Have user feedback when tiles are clicked

(cherry picked from commit 5a3bcc0)

+++ app/src/main/java/com/kpchuck/qstiles/RebootTileService.java
@@ -1,5 +1,8 @@
 package com.kpchuck.qstiles;
 
+import android.content.Intent;
+import android.widget.Toast;
+
 import eu.chainfire.libsuperuser.Shell;
 
 /**
@@ -10,6 +13,19 @@ public class RebootTileService extends android.service.quicksettings.TileService
 
     public void onClick(){
 
-        Shell.SU.run("reboot");
+        Intent it = new Intent(Intent.ACTION_CLOSE_SYSTEM_DIALOGS);
+
+        getApplicationContext().sendBroadcast(it);
+        if (Shell.SU.available()){
+            longToast("Rebooting, Please Wait...");
+              Shell.SU.run("reboot");
+        }else{
+            longToast("Ozone doesn't have root permission. Open SuperSU/magisk and try again");
+        }
+
+    }
+
+    public void longToast(String message){
+        Toast.makeText(getApplicationContext(), message, Toast.LENGTH_LONG).show();
     }
 }
